Remove the possibility to explicitly create nested Flows.
* Explicitly writing Flow[Flow[...]] will result in an error during
prechecking.

* Function assertNotNestedFlow raises the ExplicitlyNestedFlowError
exception when detecting an explicit nesting of Flows.

* It is possible to create nested flows when dealing with parametric types.
The typechecker collapses the types to ensure that we are trying to bind a
T to a Flow+[T], and not a Flow+[T] to a T.

* Added several tests in tests/encore/flow to make sure nesting is properly
handled.

* Still need to write the .out files so the tests can be part of the test
suite.

* Removed the old flowTypecheckCollapseParametric, as it was performing an
invalid test based on an incorrect reading of Flow's semantics.

* Updated test flowTypecheckSimple to remove the nested types as they are
no longer allowed.