Implement translation of async* and get*
  * There is now a spawn_star function in TaskRunner. This function works like
the spawn function, but performs an async flow call, instead of an async future
call

  * async* desugars into a call to spawn_star, just like async desugars into a
call to spawn. As such, the Task module is required to use async*

  * The internal API now has mirroring structures for everything flow related.
    * encore_flow_msg_t mirros encore_fut_msg_t, and contains a single field
(_flow) which is a pointer to the flow used in the message

    * When compiling Encore methods, the compiler will produce a new class of
C functions, ending in _flow, mirroring the _future versions of Encore methods.
These functions are called when a flow call is performed.

    * Actors messages sent in a flow context are identified with the MSG_FLOW
string, just like messages sent in a future context are identified with the
MSG_FUT string

    * The dispatch method of the classes now handle the flow messages

    * When executing the dispatch method of a class, the program will exit with
an error code of 1 if forward is used in conjunction with a flow call, as I'm
not yet sure of how to handle such a thing yet

    * Translation of async* and get* mirrors the translation of async and get.
However, one way messages and streams are not reproduced as I'm not yet sure
of how to handle them yet

    * The callMethodWithFlow function mirrors callMethodWithFuture, but maybe
a more generic solution could be used

    * Exposed the flow_result_type_t enumeration to the public, as it is
required when using flows (also renamed the structure from result_type_t as it
is now in the global namespace)

    * Rewrote small bits of the flow API, it still needs to be completed