Implement async* and get*
  * flow.c now contains a working implementation for get*
  * There are crashes related to GC when using flows, this need to be fixed
  * Added the LiftToFlow node in the AST. This node indicates that its
containing expression must be send to flow_mk_from_value
  * Fixed the coercion rules of Flows, as this caused async* to fail
typechecking when the async'ed function already returned a Flow
  * The lifting to flow could be put in a function to avoid code repetition
  * Added the async_common/async.h header file, for structure definitions
common to both the implementation of future and flow. The premake4.lua has
been updated as well
  * Added an error message when spawn / spawn_star is used (when async / async*
desugars) and that Task is not imported. This helps developper understand why
the typechecker complains about functions they didn't use
  * Renamed the acquireobject / acquireactor in the Pony runtime GC to reflect
that they are both used by the flow and future libraries, and not only by the
future library